\name{candidates.print.command}
\alias{candidates.print.command}
\title{candidates.print.command}
\description{
  Print candidate configurations (hereafter "candidates")
  as command lines.
}
\usage{
  candidates.print.command(cand, parameters)
}
\arguments{
  \item{cand}{A matrix containing the candidates (one per row).}
  \item{parameters}{A data structure similar to that provided
    by the 'readParameters' function.}
}
\value{
  None.
}
\seealso{
  \code{\link{candidates.print}} to print the candidates as a data frame.
}
\author{Manuel López-Ibáñez and Jérémie Dubois-Lacoste}
