% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParameters.R
\name{checkParameters}
\alias{checkParameters}
\title{checkParameters}
\usage{
checkParameters(parameters)
}
\arguments{
\item{parameters}{\code{ParameterSpace}\cr Data structure containing the parameter
space definition. The data structure has to similar to the one returned by the
function \code{\link{readParameters}}.}
}
\description{
FIXME: This is incomplete, for now we only repair inputs from previous irace
versions.
}
