% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser.R
\docType{data}
\name{CommandArgsParser}
\alias{CommandArgsParser}
\alias{cmdline_usage}
\title{R6 Class for parsing command-line arguments}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CommandArgsParser

cmdline_usage(cmdline_args)
}
\arguments{
\item{cmdline_args}{Definition of the command-line arguments.}
}
\description{
R6 Class for parsing command-line arguments

\code{cmdline_usage()} prints the output of \code{--help}
}
\keyword{datasets}
