\name{configurations.print.command}
\alias{configurations.print.command}
\title{configurations.print.command}
\description{
  Print configurations as command lines.
}
\usage{
  configurations.print.command(configurations, parameters)
}
\arguments{
  \item{configurations}{a data frame containing the configurations (one per row).}
  \item{parameters}{A data structure similar to that provided
    by the \code{\link{readParameters}} function.}
}
\value{
  None.
}
\seealso{
  \code{\link{configurations.print}} to print the configurations as a data frame.
}
\author{Manuel López-Ibáñez and Jérémie Dubois-Lacoste}
