% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{configurations_print_command}
\alias{configurations_print_command}
\title{Print configurations as command-line strings.}
\usage{
configurations_print_command(configurations, parameters)
}
\arguments{
\item{configurations}{\code{data.frame}\cr Parameter configurations of the
target algorithm (one per row).}

\item{parameters}{\code{ParameterSpace}\cr Data structure containing the parameter
space definition. The data structure has to similar to the one returned by the
function \code{\link{readParameters}}.}
}
\value{
None.
}
\description{
Prints configurations after converting them into a representation for the
command-line.
}
\seealso{
\code{\link[=configurations_print]{configurations_print()}} to print the configurations as a data frame.
}
\author{
Manuel López-Ibáñez and Jérémie Dubois-Lacoste
}
