% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterAnalysis.R
\name{configurationsBoxplot}
\alias{configurationsBoxplot}
\title{Creates box plots of the quality of configurations.}
\usage{
configurationsBoxplot(experiments, title = NULL,
  xlabel = "Configuration ID", ylabel = "Configuration cost",
  filename = NULL)
}
\arguments{
\item{experiments}{Matrix of performance of configurations (columns) over a set of instances (rows).}

\item{title}{(\code{NULL}) Title for the plot.}

\item{xlabel}{Label for the x axis.}

\item{ylabel}{Label for the y axis.}

\item{filename}{(\code{NULL}) Filename prefix to create a pdf file with the plot.}
}
\value{
Box plot of the performance of the configurations.
}
\description{
Creates box plots of the quality of configurations.
}
\author{
Manuel López-Ibáñez and Leslie Pérez Cáceres
}
