\name{defaultConfiguration}
\alias{defaultConfiguration}
\title{defaultConfiguration}
\description{
  'defaultConfiguration' set all parameters of \pkg{irace} (the
  so-called "configuration") to their default values.
}
\usage{
 defaultConfiguration(config = list())
}
\arguments{
  \item{config}{A list indexed by the \pkg{irace} parameter names,
  containing their values.}
}
\value{
  A list indexed by the \pkg{irace} parameter names,
  containing the default values for each parameter.
}
\details{
}
\references{
}
\seealso{
  'readConfiguration' for reading \pkg{irace} configuration.
  'printConfiguration' for printing \pkg{irace} configuration.
  'checkConfiguration' to check that a configuration is valid.
}
\examples{
}
\author{Manuel Lopez-Ibanez and Jeremie Dubois-Lacoste}
\keyword{}
