% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{irace.cmdline}
\alias{irace.cmdline}
\title{irace.cmdline}
\usage{
irace.cmdline(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{\code{commandArgs(trailingOnly = TRUE)} The arguments 
provided on the R command line as a character vector, e.g., 
\code{c("--scenario", "scenario.txt", "-p", "parameters.txt")}.
Using the  default value (not providing the parameter) is the 
easiest way to call \code{irace.cmdline}.}
}
\value{
None.
}
\description{
\code{irace.cmdline} starts \pkg{irace} using the parameters
 of the command line used to invoke R.
}
\details{
The function reads the parameters given on the command line
used to invoke R, finds the name of the scenario file,
 initializes the scenario from the file (with the function
 \link{readScenario}) and possibly from parameters passed on
 the command line. It finally starts \pkg{irace} by calling
 \link{irace.main}.
}
\seealso{
\code{\link{irace.main}} to start \pkg{irace} with a given scenario.
}
\author{
Manuel López-Ibáñez and Jérémie Dubois-Lacoste
}
