% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{irace.license}
\alias{irace.license}
\title{irace.license}
\format{An object of class \code{character} of length 1.}
\usage{
irace.license
}
\description{
A character string containing the license information of \pkg{irace}.
}
\keyword{datasets}
