% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irace2pyimp.R
\name{irace2pyimp_cmdline}
\alias{irace2pyimp_cmdline}
\title{Command-line interface to irace2pyimp}
\usage{
irace2pyimp_cmdline(argv = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{argv}{(\code{character()}) \cr Command-line arguments.}
}
\value{
None.
}
\description{
This is a command-line interface for calling the \code{\link{irace2pyimp}} function,
which converts an \code{irace.Rdata} file into the input format supported by the
parameter importance analysis tool \code{PyImp}
(https://github.com/automl/ParameterImportance). \cr The best way to use
this command line interface is to run the script \code{irace-to-pyimp}. \cr To
see usage of the script, run: irace-to-pyimp --help
}
\examples{

irace2pyimp_cmdline("--help")

}
\seealso{
\code{\link{irace2pyimp}}
}
\author{
Nguyen Dang and Manuel López-Ibáñez
}
