% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{irace_license}
\alias{irace_license}
\title{irace_license}
\format{
An object of class \code{character} of length 1.
}
\usage{
irace_license
}
\description{
A character string containing the license information of \pkg{irace}.
}
\keyword{datasets}
