% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irace_summarise.R
\name{irace_summarise}
\alias{irace_summarise}
\title{Summarise the results of a run of irace}
\usage{
irace_summarise(iraceResults)
}
\arguments{
\item{iraceResults}{\code{list()}|\code{character(1)}\cr Object created by \pkg{irace} and typically saved in the log file \code{irace.Rdata}. If a character string is given, then it is interpreted as the path to the log file from which the \code{iraceResults} object will be loaded.}
}
\value{
\code{list()}
}
\description{
Summarise the results of a run of irace
}
\examples{
irace_results <- read_logfile(system.file("exdata/irace-acotsp.Rdata",
                                          package="irace", mustWork=TRUE))
irace_summarise(irace_results)

}
\author{
Manuel López-Ibáñez
}
\concept{analysis}
