% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{irace_version}
\alias{irace_version}
\title{A character string containing the version of \code{irace} including git SHA.}
\usage{
irace_version
}
\description{
A character string containing the version of \code{irace} including git SHA.
}
