% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_rel2abs}
\alias{path_rel2abs}
\title{Converts a relative path to an absolute path. It tries really hard}
\usage{
path_rel2abs(path, cwd = getwd())
}
\arguments{
\item{path}{character string representing a relative path.}

\item{cwd}{current working directory.}
}
\value{
(`character(1)`) character string representing the absolute path
}
\description{
Converts a relative path to an absolute path. It tries really hard
}
