\name{printScenario}
\alias{printScenario}
\title{printScenario}
\description{
  'printScenario' prints the given scenario.
}
\usage{
  printScenario(scenario)
}
\arguments{
   \item{scenario}{A list where tagged elements correspond to scenario
    settings of \pkg{irace}.}
}
\seealso{
  \code{\link{readScenario}} for reading a configuration scenario from a
  file.
  \code{\link{defaultScenario}} returns the default scenario settings of
  \pkg{irace}.
  \code{\link{checkScenario}} to check that the scenario is valid.
}
\author{Manuel López-Ibáñez and Jérémie Dubois-Lacoste}
