% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablation.R
\name{read_ablogfile}
\alias{read_ablogfile}
\title{Read the log file (\code{log-ablation.Rdata}) produced by \code{\link[=ablation]{ablation()}}.}
\usage{
read_ablogfile(filename)
}
\arguments{
\item{filename}{Filename that contains the log file saved by \code{\link[=ablation]{ablation()}}. Example: \code{log-ablation.Rdata}.}
}
\value{
(\code{list()})
}
\description{
Read the log file (\code{log-ablation.Rdata}) produced by \code{\link[=ablation]{ablation()}}.
}
\concept{ablation}
