% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_logfile}
\alias{read_logfile}
\title{Read the log file produced by irace (\code{irace.Rdata}).}
\usage{
read_logfile(filename, name = "iraceResults")
}
\arguments{
\item{filename}{Filename that contains the log file saved by irace. Example: \code{irace.Rdata}.}

\item{name}{Optional argument that allows overriding the default name of the object in the file.}
}
\value{
(\code{list()})
}
\description{
Read the log file produced by irace (\code{irace.Rdata}).
}
\concept{analysis}
