% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_irace_logfile}
\alias{save_irace_logfile}
\title{Save the log generated by \pkg{{irace}} to a file (by default \code{irace.Rdata}).}
\usage{
save_irace_logfile(iraceResults, logfile)
}
\arguments{
\item{iraceResults}{\code{list()}\cr Object created by \pkg{irace} and typically saved in the log file \code{irace.Rdata}.}

\item{logfile}{\code{character(1)}\cr Filename to save \code{iraceResults}. Usually, this is given by \code{scenario$logFile}. If \code{NULL} or \code{""}, no data is saved.}
}
\description{
This function may be useful if you are manually editing the log data generated by a run of \pkg{irace}.
}
\seealso{
\code{\link[=read_logfile]{read_logfile()}}
}
\concept{analysis}
