% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser.R
\name{CommandArgsParser}
\alias{CommandArgsParser}
\alias{cmdline_usage}
\title{R6 Class for parsing command-line arguments}
\usage{
cmdline_usage(cmdline_args)
}
\arguments{
\item{cmdline_args}{Definition of the command-line arguments.}
}
\description{
R6 Class for parsing command-line arguments

R6 Class for parsing command-line arguments

\code{cmdline_usage()} prints the output of \code{--help}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CommandArgsParser$new()}}
\item \href{#method-readCmdLineParameter}{\code{CommandArgsParser$readCmdLineParameter()}}
\item \href{#method-readArg}{\code{CommandArgsParser$readArg()}}
\item \href{#method-readAll}{\code{CommandArgsParser$readAll()}}
\item \href{#method-cmdline_usage}{\code{CommandArgsParser$cmdline_usage()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandArgsParser$new(argv, argsdef)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readCmdLineParameter"></a>}}
\if{latex}{\out{\hypertarget{method-readCmdLineParameter}{}}}
\subsection{Method \code{readCmdLineParameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandArgsParser$readCmdLineParameter(paramName, default = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readArg"></a>}}
\if{latex}{\out{\hypertarget{method-readArg}{}}}
\subsection{Method \code{readArg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandArgsParser$readArg(short = "", long = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readAll"></a>}}
\if{latex}{\out{\hypertarget{method-readAll}{}}}
\subsection{Method \code{readAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandArgsParser$readAll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cmdline_usage"></a>}}
\if{latex}{\out{\hypertarget{method-cmdline_usage}{}}}
\subsection{Method \code{cmdline_usage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CommandArgsParser$cmdline_usage()}\if{html}{\out{</div>}}
}

}
}
