% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configurations.print.command}
\alias{configurations.print.command}
\title{Print configurations as command-line strings.}
\usage{
configurations.print.command(configurations, parameters)
}
\arguments{
\item{configurations}{a data frame containing the configurations (one per row).}

\item{parameters}{A data structure similar to that provided by the 
\code{\link{readParameters}} function.}
}
\value{
None.
}
\description{
Prints configurations after converting them into a representation for the
command-line.
}
\seealso{
\code{\link{configurations.print}} to print the configurations as a data frame.
}
\author{
Manuel López-Ibáñez and Jérémie Dubois-Lacoste
}
