\name{configurations.print.command}
\alias{configurations.print.command}
\title{configurations.print.command}
\description{
  Print configuration configurations (hereafter "configurations")
  as command lines.
}
\usage{
  configurations.print.command(configuration, parameters)
}
\arguments{
  \item{configuration}{A matrix containing the configurations (one per row).}
  \item{parameters}{A data structure similar to that provided
    by the 'readParameters' function.}
}
\value{
  None.
}
\seealso{
  \code{\link{configurations.print}} to print the configurations as a data frame.
}
\author{Manuel López-Ibáñez and Jérémie Dubois-Lacoste}
