% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{configurations_print}
\alias{configurations_print}
\title{Print configurations as a data frame}
\usage{
configurations_print(configurations, metadata = FALSE)
}
\arguments{
\item{configurations}{\code{data.frame}\cr Parameter configurations of the
target algorithm (one per row).}

\item{metadata}{\code{logical(1)}\cr whether to print the metadata or
not. The metadata are data for the configurations (additionally to the
value of each parameter) used by \pkg{irace}.}
}
\value{
None.
}
\description{
Print configurations as a data frame
}
\seealso{
\code{\link[=configurations_print_command]{configurations_print_command()}} to print the configurations as command-line strings.
}
\author{
Manuel López-Ibáñez and Jérémie Dubois-Lacoste
}
