% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterAnalysis.R
\name{getConfigurationById}
\alias{getConfigurationById}
\title{Returns the configurations selected by ID.}
\usage{
getConfigurationById(iraceResults, ids, drop.metadata = FALSE)
}
\arguments{
\item{iraceResults}{(\code{list()}|\code{character(1)}) \code{iraceResults} object created by \pkg{irace} and typically saved in the log file \code{irace.Rdata}. If a character string is given, then it is interpreted as the path to the log file from which the \code{iraceResults} object will be loaded.}

\item{ids}{The id or a vector of ids of the candidates configurations to obtain.}

\item{drop.metadata}{Remove metadata, such the configuration ID and
the ID of the parent, from the returned configurations. See
\code{\link[=removeConfigurationsMetaData]{removeConfigurationsMetaData()}}.}
}
\value{
A data frame containing the elite configurations required.
}
\description{
Returns the configurations selected by ID.
}
\author{
Manuel López-Ibáñez and Leslie Pérez Cáceres
}
\concept{analysis}
