% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterAnalysis.R
\name{getConfigurationByIteration}
\alias{getConfigurationByIteration}
\title{Returns the configurations by the iteration in which they were executed.}
\usage{
getConfigurationByIteration(iraceResults, iterations, drop.metadata = FALSE)
}
\arguments{
\item{iraceResults}{\code{list()}|\code{character(1)}\cr Object created by \pkg{irace} and typically saved in the log file \code{irace.Rdata}. If a character string is given, then it is interpreted as the path to the log file from which the \code{iraceResults} object will be loaded.}

\item{iterations}{(\code{integer()})\cr The iteration number or a vector of iteration numbers from where
the configurations should be obtained. Negative values start counting from the last iteration.}

\item{drop.metadata}{\code{logical(1)}\cr Remove metadata, such as the
configuration ID and the ID of the parent, from the returned
configurations.  See \code{\link[=removeConfigurationsMetaData]{removeConfigurationsMetaData()}}.}
}
\value{
A data frame containing the elite configurations required.
}
\description{
Returns the configurations by the iteration in which they were executed.
}
\examples{
log_file <- system.file("exdata/irace-acotsp.Rdata", package="irace", mustWork=TRUE)
getConfigurationByIteration(log_file, iterations = c(-2, -1), drop.metadata = TRUE)

}
\author{
Manuel López-Ibáñez and Leslie Pérez Cáceres
}
\concept{analysis}
