% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterAnalysis.R
\name{getConfigurationByIteration}
\alias{getConfigurationByIteration}
\title{Returns the configurations by the iteration in which they were executed.}
\usage{
getConfigurationByIteration(iraceResults = NULL, logFile = NULL,
  iterations, drop.metadata = FALSE)
}
\arguments{
\item{iraceResults}{(\code{NULL}) Object created by \pkg{irace} and saved in \code{scenario$logFile}.}

\item{logFile}{(\code{NULL}) Log file created by \pkg{irace}, this file must contain the 
\code{iraceResults} object.}

\item{iterations}{The iteration number or a vector of iteration numbers from where 
the configurations should be obtained.}

\item{drop.metadata}{(\code{FALSE}) Remove metadata, such the configuration ID and
the ID of the parent, from the returned configurations. See
\code{\link{removeConfigurationsMetaData}}.}
}
\value{
A data frame containing the elite configurations required.
}
\description{
Returns the configurations by the iteration in which they were executed.
}
\author{
Manuel López-Ibáñez and Leslie Pérez Cáceres
}
