% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterAnalysis.R
\name{getFinalElites}
\alias{getFinalElites}
\title{Return the elite configurations of the final iteration.}
\usage{
getFinalElites(iraceResults = NULL, logFile = NULL, n = 0,
  drop.metadata = FALSE)
}
\arguments{
\item{iraceResults}{Object created by \pkg{irace} and saved in \code{scenario$logFile}.}

\item{logFile}{Log file created by \pkg{irace}, this file must contain the 
\code{iraceResults} object.}

\item{n}{Number of elite configurations to return, if \code{n} is larger than the 
number of configurations, then only the existing ones are returned.}

\item{drop.metadata}{Remove metadata, such the configuration ID and
the ID of the parent, from the returned configurations. See
\code{\link{removeConfigurationsMetaData}}.}
}
\value{
A data frame containing the elite configurations required.
}
\description{
Return the elite configurations of the final iteration.
}
\author{
Manuel López-Ibáñez and Leslie Pérez Cáceres
}
