\name{irace.usage}
\alias{irace.usage}
\title{irace.usage}
\description{
  iraceUsage' prints all possible arguments for the configuration
  to be used by \pkg{irace}, with the corresponding switches and
  a short description.
}
\usage{
  irace.usage()
}
\arguments{
  None.
}
\value{
  None.
}
\details{
}
\references{}
\seealso{
}
\examples{}
\author{Manuel Lopez-Ibanez and Jeremie Dubois-Lacoste}
\keyword{}
