\name{irace.usage}
\alias{irace.usage}
\title{irace.usage}
\description{
  This function prints all possible arguments for the configuration
  to be used by \pkg{irace}, with the corresponding switches and
  a short description.
}
\usage{
  irace.usage()
}
\author{Manuel López-Ibáñez and Jérémie Dubois-Lacoste}
