% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{irace.usage}
\alias{irace.usage}
\title{irace.usage}
\usage{
irace.usage()
}
\description{
\code{irace.usage}  This function prints all command-line options of \pkg{irace},
  with the corresponding switches and a short description.
}
\author{
Manuel López-Ibáñez and Jérémie Dubois-Lacoste
}
