% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\docType{data}
\name{irace.version}
\alias{irace.version}
\title{irace.version}
\format{An object of class \code{character} of length 1.}
\usage{
irace.version
}
\description{
A character string containing the version of \pkg{irace}.
}
\keyword{datasets}
