% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterAnalysis.R
\name{parallelCoordinatesPlot}
\alias{parallelCoordinatesPlot}
\title{parallelCoordinatesPlot}
\usage{
parallelCoordinatesPlot(configurations, parameters,
  param_names = parameters$names, hierarchy = TRUE, filename = NULL,
  pdf.width = 14, mar = c(8, 1, 4, 1))
}
\arguments{
\item{configurations}{(\code{data.frame}) \cr Parameter configurations of the
target algorithm (one per row).}

\item{parameters}{(\code{list()}) \cr Data structure containing the parameter
space definition. The data structure has to similar to the one returned by the
function \code{\link{readParameters}}.}

\item{param_names}{Parameters names that should be included. Default: parameters$names.}

\item{hierarchy}{If \code{TRUE} conditional parameters will be displayed in a different 
plot. Default \code{TRUE}.}

\item{filename}{Filename prefix to generate the plots. If \code{NULL} the plot 
displayed but not saved.}

\item{pdf.width}{Width for the pdf file generated.}

\item{mar}{Margin to use for the plot. See \code{\link{par}}.}
}
\value{
A set of parallel coordinates plots showing the parameters values. 
  If a filename is provided this plots are saved in one or more files.
}
\description{
\code{parallelCoordinatesPlot}  plots a set of parameter configurations in 
  parallel coordinates.
}
\examples{
\donttest{
 ## To use data obtained by irace
 # First, load the data produced by irace.
 irace.logfile <- file.path(system.file(package="irace"), "exdata", "irace-acotsp.Rdata")
 load(irace.logfile)
 attach(iraceResults)
 parallelCoordinatesPlot(allConfigurations, parameters, hierarchy = FALSE)
}

}
\seealso{
\code{\link{readParameters}} to obtain a valid parameter structure from a parameters file.
 \code{\link{readConfigurationsFile}} to obtain a set of target algorithm configurations from 
   a configurations file.
}
\author{
Manuel López-Ibáñez and Leslie Pérez Cáceres
}
