% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablation.R
\name{plotAblation}
\alias{plotAblation}
\title{Create plot from an ablation log}
\usage{
plotAblation(ab.log = NULL, abLogFile = NULL, pdf.file = NULL,
  pdf.width = 20, type = c("mean", "boxplot"), mar = par("mar"),
  ylab = "Mean configuration cost", ...)
}
\arguments{
\item{ab.log}{Ablation log returned by \code{\link{ablation}}.}

\item{abLogFile}{Rdata file containing the ablation log.}

\item{pdf.file}{Output filename.}

\item{pdf.width}{Width provided to create the pdf file.}

\item{type}{Type of plots. Supported values are \code{"mean"} and
\code{"boxplot"}.}

\item{mar}{Vector with the margins for the ablation plot.}

\item{ylab}{Label of y-axis.}

\item{...}{Further graphical parameters may also be supplied as
arguments. See \code{plot.default}.}
}
\description{
Create plot from an ablation log
}
\seealso{
\code{\link{ablation}}
}
\author{
Leslie Pérez Cáceres and Manuel López-Ibáñez
}
