\name{printConfiguration}
\alias{printConfiguration}
\title{printConfiguration}
\description{
  'printConfiguration' prints the internal configuration of 
  \pkg{irace}.
}
\usage{
  printConfiguration(configuration)
}
\arguments{
  \item{configuration}{A list where tagged elements correspond to configuration
    parameters of \pkg{irace}.}
}
\seealso{
  \code{\link{readConfiguration}} for reading the \pkg{irace}
  configuration from a file.
  \code{\link{defaultConfiguration}} to get the default configuration.
  \code{\link{checkConfiguration}} to check that the configuration is valid.
}
\author{Manuel López-Ibáñez and Jérémie Dubois-Lacoste}
