% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfiguration.R
\name{printScenario}
\alias{printScenario}
\title{Prints the given scenario}
\usage{
printScenario(scenario)
}
\arguments{
\item{scenario}{A list where tagged elements correspond to scenario
settings of \pkg{irace}.}
}
\description{
Prints the given scenario
}
\seealso{
\describe{
 \item{\code{\link{readScenario}}}{for reading a configuration scenario from a file.}
 \item{\code{\link{printScenario}}}{prints the given scenario.}
 \item{\code{\link{defaultScenario}}}{returns the default scenario settings of \pkg{irace}.}
 \item{\code{\link{checkScenario}}}{to check that the scenario is valid.}
}
}
\author{
Manuel López-Ibáñez and Jérémie Dubois-Lacoste
}
