% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParameters.R
\name{read_pcs_file}
\alias{read_pcs_file}
\title{read_pcs_file}
\usage{
read_pcs_file(file, digits = 4, debugLevel = 0, text)
}
\arguments{
\item{file}{(\code{character(1)}) \cr Filename containing the definitions of
the parameters to be tuned.}

\item{digits}{The number of decimal places to be considered for the real
parameters.}

\item{debugLevel}{(\code{integer(1)}) \cr Larger values produce more verbose output.}

\item{text}{(\code{character(1)}) \cr If \code{file} is not supplied and this is,
then parameters are read from the value of \code{text} via a text connection.}
}
\value{
A string representing the parameters in irace format.
}
\description{
Read parameters in PCS (AClib) format and write them in irace format.
}
\details{
Either \code{file} or \code{text} must be given. If \code{file} is given, the
parameters are read from the file \code{file}. If \code{text} is given instead,
the parameters are read directly from the \code{text} character string.
In both cases, the parameters must be given (in \code{text} or in the file
whose name is \code{file}) in the expected form.  See the documentation
for details.  If none of these parameters is given, \pkg{irace}
will stop with an error.
}
\examples{
 ## Read the parameters directly from text
 pcs.table <- '
 # name       values               [conditions (using R syntax)]
 algorithm    {as,mmas,eas,ras,acs}[as]
 localsearch  {0, 1, 2, 3}[0]
 alpha        [0.00, 5.00][1]
 beta         [0.00, 10.00][1]
 rho          [0.01, 1.00][0.95]
 ants         [5, 100][10]i
 q0           [0.0, 1.0][0]
 rasrank      [1, 100][1]i
 elitistants  [1, 750][1]i
 nnls         [5, 50][5]i
 dlb          {0, 1}[1] 
 Conditionals:
 q0 | algorithm in {acs}
 rasrank | algorithm in {ras}
 elitistants | algorithm in {eas}
 nnls | localsearch in {1,2,3}
 dlb | localsearch in {1,2,3}
 '
 parameters.table <- read_pcs_file(text=pcs.table)
 parameters <- readParameters(text=parameters.table)
 str(parameters)

}
\author{
Manuel López-Ibáñez
}
