% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParameters.R
\name{readParametersData}
\alias{readParametersData}
\title{Reads the parameters to be tuned by \pkg{irace} from a r data structure.}
\usage{
readParametersData(data, n)
}
\arguments{
\item{data}{A list containing the definitions of the parameters. THis list is
structured as follows:
\describe{
\item{\code{names}}{Vector that contains the names of the parameters.}
\item{\code{types}}{Vector that contains the type of each parameter 'i', 'c', 'r', 'o'.
Numerical parameters can be sampled in a log-scale with 'i,log' and 'r,log'
(no spaces).}
\item{\code{switches}}{Vector that contains the switches to be used for the
parameters on the command line.}
\item{\code{domain}}{List of vectors, where each vector may contain two
values (minimum, maximum) for real and integer parameters, or
possibly more for categorical parameters.}
\item{\code{conditions}}{List of R logical expressions, with variables
corresponding to parameter names.}
}}

\item{n}{The number of parameters as an integer (the function will not automatically
type cast it).}
}
\value{
A list containing the definitions of the parameters read. The list is
structured as follows:
\describe{
\item{\code{names}}{Vector that contains the names of the parameters.}
\item{\code{types}}{Vector that contains the type of each parameter 'i', 'c', 'r', 'o'.
Numerical parameters can be sampled in a log-scale with 'i,log' and 'r,log'
(no spaces).}
\item{\code{switches}}{Vector that contains the switches to be used for the
parameters on the command line.}
\item{\code{domain}}{List of vectors, where each vector may contain two
values (minimum, maximum) for real and integer parameters, or
possibly more for categorical parameters.}
\item{\code{conditions}}{List of R logical expressions, with variables
corresponding to parameter names.}
\item{\code{isFixed}}{Logical vector that specifies which parameter is fixed
and, thus, it does not need to be tuned.}
\item{\code{nbParameters}}{An integer, the total number of parameters.}
\item{\code{nbFixed}}{An integer, the number of parameters with a fixed value.}
\item{\code{nbVariable}}{Number of variable (to be tuned) parameters.}
\item{\code{depends}}{List of character vectors, each vector specifies
which parameters depend on this one.}
\item{\code{isDependent}}{Logical vector that specifies which parameter has
a dependent domain.}
}
}
\description{
Reads the parameters to be tuned by \pkg{irace} from a r data structure.
}
\details{
This function is designed to used as an API for binding with other languages
such as python with iracepy. As such, there is no input validation
(validations should be by the packages using it, such as iracepy). Invalid
input such as different lengths for the vectors and list will result in undefined behavior.

A fixed parameter is a parameter that should not be sampled but
instead should be always set to the only value of its domain.  In this
function we set isFixed to TRUE only if the parameter is a categorical
and has only one possible value.  If it is an integer and the minimum
and maximum are equal, or it is a real and the minimum and maximum
values satisfy \code{round(minimum, digits) == round(maximum, digits)},
then the parameter description is rejected as invalid to identify
potential user errors.
}
\examples{
s <- '
initial_temp       "" r,log (0.02, 5e4)
restart_temp_ratio "" r,log (1e-5, 1)
visit              "" r     (1.001, 3)
b                  "" r     ("visit", "visit + 10")
accept             "" r     (-1e3, -5) | no_local_search == "ccc"
no_local_search    "" o     ("","1","ccc") 
'


d = list()
d$names <- c(
    'initial_temp',
    'restart_temp_ratio',
    'visit',
    'b',
    'accept',
    'no_local_search'
)

d$switches <- c(
    '',
    '',
    '',
    '',
    '',
    ''
)

d$types <- c(
    'r,log',
    'r,log',
    'r',
    'r',
    'r',
    'o'
)

d$conditions <- list(
    TRUE,
    TRUE,
    TRUE,
    TRUE,
    expression(no_local_search == "ccc"),
    TRUE
)

d$domains <- list(
    c(0.02, 5e4),
    c(1e-5, 1),
    c(1.001, 3),
    expression(visit, visit + 10),
    c(-1e3, -5),
    c('', '1', 'ccc')
)



t <- readParameters(text = s, digits = 5)

u <- readParametersData(data = d, n = 6L)

identical(t, u)
# The two methods of reading parameters give identical results

}
\author{
Deyao Chen
}
