% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeConfigurationsMetaData}
\alias{removeConfigurationsMetaData}
\title{removeConfigurationsMetaData}
\usage{
removeConfigurationsMetaData(configurations)
}
\arguments{
\item{configurations}{(\code{data.frame}) \cr Parameter configurations of the
target algorithm (one per row).}
}
\value{
The same matrix without the "metadata".
}
\description{
Remove the columns with "metadata" of a matrix containing some
configuration configurations. These "metadata" are used internaly
by \pkg{irace}. This function can be used e.g. before printing
the configurations, to output only the values for the parameters
of the configuration without data possibly useless to the user.
}
\seealso{
\code{\link{configurations.print.command}} to print the configurations as command lines.
  \code{\link{configurations.print}} to print the configurations as a data frame.
}
\author{
Manuel López-Ibáñez and Jérémie Dubois-Lacoste
}
