% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{testing_fromfile}
\alias{testing_fromfile}
\title{Test configurations given an explicit table of configurations and a scenario file}
\usage{
testing_fromfile(filename, scenario)
}
\arguments{
\item{filename}{\code{character(1)}\cr Path to a file containing configurations: one configuration
per line, one parameter per column, parameter names in header.}

\item{scenario}{\code{list()}\cr Data structure containing \pkg{irace}
settings. The data structure has to be the one returned by the function
\code{\link[=defaultScenario]{defaultScenario()}} or \code{\link[=readScenario]{readScenario()}}.}
}
\value{
iraceResults
}
\description{
Executes the testing of an explicit list of configurations given in
\code{filename} (same format as in \code{\link[=readConfigurationsFile]{readConfigurationsFile()}}). A \code{logFile} is
created unless disabled in \code{scenario}. This may overwrite an existing one!
}
\seealso{
\code{\link[=testing_fromlog]{testing_fromlog()}} provides a different interface for testing.
}
\author{
Manuel López-Ibáñez
}
\concept{running}
